requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-caching==2.1.0

langchain==0.1.14
langchain-openai==0.1.3
langchain-google-genai==0.0.6\

openai==1.55.3
google-generativeai==0.3.2

pypdf==4.2.0
tiktoken==0.6.0
faiss-cpu==1.8.0

celery==5.3.6
kombu==5.3.5
psycopg2-binary==2.9.9

nltk==3.8.1
textblob==0.18.0.post0

pydantic==1.10.13
python-dotenv
gunicorn
sqlalchemy
